requirements.txt
Flask==3.0.3
Pillow==11.1.0
cryptography==42.0.5
